fix: static building
