+++ next.config.js
@@ -1,5 +1,6 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
+  // output: "export", // Commented out to allow dynamic routes with locale
   images: {
     domains: ['localhost'],
   },
