+++ package.json
@@ -7,6 +7,7 @@
     "dev": "next dev",
     "lint": "next lint",
     "start": "next start",
+    "export": "next export",
     "backend": "node backend/server.js",
     "backend:dev": "nodemon backend/server.js"
   },
